wp-ffv-banner: Directly use dsgvo-youtube plugin
The dsgvo-youtube can take care of generating the correct HTML code for the
inner part of the banner. Only some styles have to be adjusted to make sure
that the height is limited (while the content is centered) and the width of
the image is the full with of the page.

Signed-off-by: Sven Eckelmann <sven@narfation.org>